+++ PSql/_Commands/ExpandSqlCmdDirectivesCommand.cs
@@ -16,6 +16,7 @@
 
 using System.Collections;
 using System.Management.Automation;
+using Prequel;
 
 namespace PSql;