+++ PSql/_Commands/InvokeSqlCommand.cs
@@ -17,6 +17,7 @@
 using System.Collections;
 using System.Data;
 using System.Management.Automation;
+using Prequel;
 
 namespace PSql;