+++ PSql/_Utilities/SqlCmdPreprocessorExtensions.cs
@@ -16,6 +16,7 @@
 
 using System.Collections;
 using System.Globalization;
+using Prequel;
 
 namespace PSql;