+++ Make.ps1
@@ -5,8 +5,8 @@
 .DESCRIPTION
     This script is similar to a makefile.
 
-    Copyright 2023 Subatomix Research Inc.
-    SPDX-License-Identifier: ISC
+    Copyright Subatomix Research Inc.
+    SPDX-License-Identifier: MIT
 #>
 [CmdletBinding(DefaultParameterSetName="Test")]
 param (