+++ PSql.Tests/Tests.Unit/CmdletSqlMessageLoggerTests.cs
@@ -1,5 +1,5 @@
-// Copyright 2023 Subatomix Research Inc.
-// SPDX-License-Identifier: ISC
+// Copyright Subatomix Research Inc.
+// SPDX-License-Identifier: MIT
 
 using System.Management.Automation.Internal;
 using System.Reflection;