+++ CHANGES.md
@@ -5,6 +5,7 @@ Most lines should begin with one of these words:
 *Add*, *Fix*, *Update*, *Change*, *Deprecate*, *Remove*.
 
 ## [Unreleased](https://github.com/sharpjs/PSql/compare/release/2.1.0..HEAD)
+- Change target to PowerShell 7.2 / .NET 6.
 - Change how the module loads dependencies.  Now, the module loads dependencies
   into a private context using [the recommended technique](https://learn.microsoft.com/en-us/powershell/scripting/dev-cross-plat/resolving-dependency-conflicts)
   to prevent errors if other moduels load conflicting dependencies.