+++ CHANGES.md
@@ -11,7 +11,7 @@ Most lines should begin with one of these words:
   to prevent errors if other moduels load conflicting dependencies.
 - Remove `Cmdlet`, moving its `WriteHost` method to an extension method.
 - Remove `IConsole`, replacing it with a simpler `ISqlMessageLogger`.
-- Update Microsoft.Data.SqlClient to [5.1.5](https://github.com/dotnet/SqlClient/blob/main/release-notes/5.1/5.1.5.md)
+- Update Microsoft.Data.SqlClient to [5.2.2](https://github.com/dotnet/SqlClient/blob/main/release-notes/5.2/5.2.2.md)
 - Add retries on transient failures.
 - Fix `EncryptionMode.Default` requiring certificate validation for local
   connections.